requirements.txt
google-generativeai>=0.5.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-cloud-vision>=3.4.4
pydantic>=2.0.0
streamlit==1.32.2
pandas==2.2.1
plotly==5.20.0
PyMuPDF>=1.23.0
google-genai>=0.3.0
Pillow>=10.0.0
